+++ publish/CHANGELOG.md
@@ -1,5 +1,8 @@
 - v1.10
+  - Adds a new field `pokeDelay` to delay pokes.
+  - Adds a new field `rpcDelay` to delay RPC calls.
   - Adds a new field `customRpc` to trigger any RPC call.
+  - Changes the field `delay` to `spawnDelay` to make it more explicit.
   - Breaking change: Poke can now also poke the poker.
 
 - v1.9